+++ py_css_styleguide/__init__.py
@@ -1,2 +1,2 @@
 """CSS Manifest driven styleguide for your project"""
-__version__ = '0.2.0-pre.3'
+__version__ = '0.2.0-pre.4'